+++ packages/flame_bloc/example/lib/src/inventory/bloc/inventory_event.dart
@@ -1,0 +1,21 @@
+part of 'inventory_bloc.dart';
+
+abstract class InventoryEvent extends Equatable {
+  const InventoryEvent();
+}
+
+class WeaponEquiped extends InventoryEvent {
+  final Weapon weapon;
+
+  const WeaponEquiped(this.weapon);
+
+  @override
+  List<Object?> get props => [weapon];
+}
+
+class NextWeaponEquiped extends InventoryEvent {
+  const NextWeaponEquiped();
+
+  @override
+  List<Object?> get props => [];
+}