+++ packages/flame_bloc/lib/flame_bloc.dart
@@ -1,0 +1,1 @@
+export 'src/flame_bloc_game.dart';