+++ packages/flame/CHANGELOG.md
@@ -9,7 +9,8 @@
  - Make super.onLoad mandatory to avoid memory leaks
  - `QueryableOrderedSet`'s `strictMode` is configurable so it is no longer necessary to call `register` before `query`
  - Add option to rotate `SpriteWidget`
- - Fix bug where onRemove was called during resizing
+ - Fix bug where `onRemove` was called during resizing
+ - Add `onAttach` and `onDetach` to `Game`
 
 ## [1.0.0-releasecandidate.14]
  - Reset effects after they are done so that they can be repeated